[WGSL] Override initializers should not be serialized https://bugs.webkit.org/show_bug.cgi?id=262385 rdar://116246863
Reviewed by Mike Wyrzykowski.

A one line fix to unbreak the shadowMapping WebGPU sample. In order to implement this
properly we also need to expose the default value to the API, which gets a little more
complicated since the default value can be an expression that depends on other values,
so for now this is enough to get the demo working, and we'll implement the rest in a
follow-up patch.

* Source/WebGPU/WGSL/Metal/MetalFunctionWriter.cpp:
(WGSL::Metal::FunctionDefinitionWriter::serializeVariable):

Canonical link: https://commits.webkit.org/268718@main