[WPE] Timers are not being scheduled when animations take all the CPU https://bugs.webkit.org/show_bug.cgi?id=262477
Reviewed by NOBODY (OOPS!).

The ThreadedDisplayRefreshMonitor can take time (layout, ...) and has an higher priority
than the main SharedTimer.
This means that animations could prevent the timers from executing, if the JS relies on this
to stop the animation the page appears to be stuck.

This is the same as https://commits.webkit.org/226833@main done for GTK.

* Source/WTF/wtf/glib/RunLoopSourcePriority.h: