REGRESSION(268723@main): Broke pyOpenSSL & Twisted https://bugs.webkit.org/show_bug.cgi?id=262512 rdar://problem/116373366
Reviewed by Jonathan Bedard.

pyOpenSSL requires on cryptography>=38.0.0,<42,!=40.0.0,!=40.0.1,
whereas we AutoInstall 36.0.2. It's unclear how this works on the WPE
bots, but I leave that as an exercise for those who maintain those bots.

Twisted 23.8.0 cannot be installed by the AutoInstaller, because we
don't support PEP 517 builders. Again, unclear how this works on the WPE
bots.

This does in principle leave this broken on all configurations on Python
3.11 or later, but should hopefully avoid re-breaking the WPE bots.

* Tools/Scripts/webkitpy/autoinstalled/twisted.py:

Canonical link: https://commits.webkit.org/268757@main