animations/animation-delay-changed.html is randomly failing https://bugs.webkit.org/show_bug.cgi?id=111612
Reviewed by Don Olmstead.

In this test case, both box1 and box2 move from 0 to 200, and go back
to 0 when the animations finish. The test case checks if box1 position
is around 200. However, it's the end of the animation. Thus, the
position randomly results in around 200 or 0.

On the other hand, the test case also checks if box2 position is
around 100. It's the middle of the animation. Do the same for box1.
Check the position of box1 at the middle of the animation.

* LayoutTests/animations/animation-delay-changed-expected.txt:
* LayoutTests/animations/animation-delay-changed.html:
* LayoutTests/platform/glib/TestExpectations:
* LayoutTests/platform/ios/TestExpectations:
* LayoutTests/platform/mac/TestExpectations:
* LayoutTests/platform/wincairo/TestExpectations:

Canonical link: https://commits.webkit.org/268812@main