[ANGLE] Fix update-angle after 268837@main https://bugs.webkit.org/show_bug.cgi?id=262671 rdar://116496840
Unreviewed build fix.

The use of grep to filter the output of git status interacts poorly with set
-e. When grep doesn't match any lines, a status code of 1 is returns with causes
the script to immediately exit when set -e is used.

Since we use the output, not the status code, to determine whether to proceed,
swallow the error by forcing the command result to true.

* Tools/Scripts/update-angle:

Canonical link: https://commits.webkit.org/268890@main