An Array index in CloneSerializer and CloneDeserializer can be confused for NonIndexPropertiesTag. https://bugs.webkit.org/show_bug.cgi?id=262616 rdar://116034413
Reviewed by Keith Miller, Sihui Liu and Chris Dumez.

CloneSerializer and CloneDeserializer were previously using NonIndexPropertiesTag as the terminator of
the indexed property section of an Array.  However, NonIndexPropertiesTag's encoding is 0xFFFFFFFD,
which is less than MAX_ARRAY_INDEX (0xFFFFFFFE) i.e. an index of 0xFFFFFFFD can be confused for the
NonIndexPropertiesTag, resulting type confusion.

This patch changes the structure of a serialized Array to always terminate its indexed property section
with a TerminatorTag (0xFFFFFFFF) first before looking for either a NonIndexPropertiesTag or another
TerminatorTag.  The presence of a NonIndexPropertiesTag after the 1st TerminatorTag indicates the
presence of a non-indexed properties section.  The presense of a TerminatorTag immediately after the
1st TerminatorTag indicates that the non-indexed properties section is empty.

Also updated the comment describing the shape of a serialized Array, and rebased a test.

* LayoutTests/fast/storage/serialized-script-value.html:
* Source/WebCore/bindings/js/SerializedScriptValue.cpp:
(WebCore::CloneSerializer::serialize):
(WebCore::CloneDeserializer::deserialize):

Canonical link: https://commits.webkit.org/267815.202@safari-7617-branch