NEW TEST(265708@main): [ wk2 ] fast/mediastream/video-srcObject-set-twice.html is a flaky ImageOnlyFailure. https://bugs.webkit.org/show_bug.cgi?id=261920 rdar://115868374
Reviewed by Jean-Yves Avenard.

The flakiness comes from the fact that the canvas capture video frame is not always properly rendered.
While we should fix this, we can make the test no longer flaky by relying on the existing take snapshot test.
This keeps the current coverage of the test which is to check that setting twice srcObject will still allow rendering video.

* LayoutTests/fast/mediastream/video-srcObject-set-twice-expected.html: Removed.
* LayoutTests/fast/mediastream/video-srcObject-set-twice-expected.txt: Added.
* LayoutTests/fast/mediastream/video-srcObject-set-twice.html:
* LayoutTests/platform/wk2/TestExpectations:

Canonical link: https://commits.webkit.org/268908@main