Data Isolation/PSON bypass due to UI-side PageLoadState state-machine relying on data which is distinct from that used to make Policy/Network Load decisions https://bugs.webkit.org/show_bug.cgi?id=257732 rdar://107186055
Reviewed by Chris Dumez.

When `didStartProvisionalLoadForFrame` is called, pageLoadState is updated to store the provisional URL
with a value passed from the web process. This URL is later consulted in `processForNavigationInternal`
when determining if the navigation is same-site. Since this URL is coming from the web process, we
should verify that the URL has not been changed from when it was set on the navigation object in
`decidePolicyForNavigationAction`.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::didStartProvisionalLoadForFrameShared):

Originally-landed-as: 265870.10@safari-7616-branch (b5aa6d4342b7). rdar://116424252
Canonical link: https://commits.webkit.org/269100@main