Regression (266514@main) Pages keep crashing when certain extensions are enabled https://bugs.webkit.org/show_bug.cgi?id=262897 rdar://116489437
Reviewed by Alex Christensen.

If the WebProcess returns a GetFrameInfo IPC for a frame that no longer exists
(e.g. it just got removed from the document), we would send a default-constructed
FrameInfoData object to the UIProcess. This used to work fine until 266514@main
made FrameInfoData.frameID non-optional. As a result, we would fail decoding of
the FrameInfoData in the UIProcess due to the missing `frameID` and terminate the
WebProcess.

To address the issue, I am now having the GetFrameInfo IPC returns a
std::optional<FrameInfoData>, so that we can return std::nullopt when the frame
in question doesn't exist.

* Source/WebKit/UIProcess/WebFrameProxy.cpp:
(WebKit::WebFrameProxy::getFrameInfo):
* Source/WebKit/WebProcess/WebPage/WebPage.cpp:
(WebKit::WebPage::getFrameInfo):
* Source/WebKit/WebProcess/WebPage/WebPage.h:
* Source/WebKit/WebProcess/WebPage/WebPage.messages.in:

Canonical link: https://commits.webkit.org/269118@main