[EWS] Add WPE API testers for the EWS https://bugs.webkit.org/show_bug.cgi?id=262478
Reviewed by Aakash Jain and Jonathan Bedard.

Add a new test queue to the EWS to run API tests
with the WPE port. It adds 4 new workers for this
new test queue like we already have for the GTK port.

This patch only adds the new testers on the EWS,
the changes on the UI to add the new status bubbles
will come on a follow-up patch later once we verify
that the new queue is running as expected.

* Tools/CISupport/ews-build/config.json:
* Tools/CISupport/ews-build/factories_unittest.py:
(TestExpectedBuildSteps):

Canonical link: https://commits.webkit.org/269162@main