_WKWebExtensionMatchPattern matchPatternWithString should fail if passed a string with a port. https://webkit.org/b/262976 rdar://problem/116760000
Reviewed by Brian Weinstein.

Return InvalidHost error if a pattern is passed that has a port. Ports are not valid in Chrome and
Firefox, and the UserContentURLPattern class didn't check them either. It was happily putting the
port in the host string and failing to match later.

Also return InvalidHost if the host has "@", which is used for user login info.

* Source/WebCore/page/UserContentURLPattern.cpp:
(WebCore::UserContentURLPattern::UserContentURLPattern):
(WebCore::UserContentURLPattern::parse):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/WKWebExtensionMatchPattern.mm:
(TestWebKitAPI::TEST):

Canonical link: https://commits.webkit.org/269182@main